requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yt-dlp>=2025.1.26
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.6

# OpenAI SDK (Whisper API, TTS API, Chat API)
openai>=1.12.0

# Fallback модели (опционально, для работы без API)
openai-whisper>=20231117
edge-tts>=6.1.0
transformers>=4.37.0
sentencepiece>=0.1.99
torch>=2.0.0

# Note: ffmpeg must be installed on the system
# macOS: brew install ffmpeg
# Ubuntu: sudo apt install ffmpeg
# Windows: download from https://ffmpeg.org/download.html
